backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP & Networking
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# HTML Parsing & Web Scraping
beautifulsoup4==4.12.2
lxml==4.9.3

# DNS & Network Tools
dnspython==2.4.2
python-nmap==0.7.1
python-whois==0.8.0

# External APIs
shodan==1.31.0

# Rate Limiting
slowapi==0.1.9

# Logging & Monitoring
structlog==23.2.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Production Server
gunicorn==21.2.0

# Environment Configuration
python-dotenv==1.0.0

# Cryptography
cryptography==41.0.7

# Date/Time Handling
python-dateutil==2.8.2

# File Processing
Pillow==10.1.0

# Async Support
asyncio==3.4.3